Fixed mac songcast bug with startup network adapters:
1. Songcast configured to en1 on 10.2.0.0 subnet
2. Close songcast
3. Change wireless en1 to be on different network 10.15.1.0
4. Start songcast - the previously selected network is not found and no devices are discovered.

The previously selected network does not appear in the network adapter list and it appears as though the 10.15.1.0 subnet is selected when it really isn't.
Selecting the new 10.15.1.0 network gets things going.

Change was made to fix this. The old selected network now appears in the list as the selected network - with "Network adapter not present" - so at least the user
can now see that they need to change the subnet.